refactoring the {tensor,dual,tuple} headers
